requirements.txt
APScheduler==3.6.3
async-timeout==4.0.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.2.2
huey==2.4.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
pymongo==4.3.3
python-decouple==3.6
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==4.4.0
six==1.16.0
tornado==6.1
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
Werkzeug==2.2.2
zipp==3.11.0
